correcao_03-oop-classes

+++ assignments/01-programming/03-oop-classes/11-parking/submissions/luiz-eduardo-s-carneiro/src/ParkingGarage.java
@@ -5,7 +5,7 @@ public class ParkingGarage{
     public static double calculateCharges(double hours){
         double taxa = 2;
         if(hours > 3){
-            double extraTime = (int)(hours - 3); 
+            double extraTime = hours - 3; 
             taxa = taxa + extraTime * 0.5;
         } 
         if(taxa > 10){
